Add methods to fetch and display user profile details
Fetches and displays the information given by the user. Each function checks if the data exists and displays it accordingly. If no data is available, a default message is shown.

Reason I have been using classes is because it optimizes the functionality of the profile aspect of the file, It makes it easier for me to organize anything within the user profiles.

I also plan on adding a method to ProfileInfo to query suggested matches from the database and extend ProfileInfoView to display them. My assumption is that these classes will make potential features a lot easier to develop.